FEAT -> Validação grades (0 - 5)

+++ hackweek-backend/dtos/RatingCriterionDTO.cs
@@ -4,7 +4,7 @@ namespace hackweek_backend.dtos
 {
     public class RatingCriterionDTO
     {
-        public int Grade { set; get; }
+        public double Grade { set; get; }
         public int CriterionId { get; set; }
     }
 }
